[CHORE] Rename crystal_rock_1.webp to ascension_crystal.webp
